Separate store from app to require it on server side
